fix flow Debug
