+++ internal/sockets/grpc_request.go
@@ -1,0 +1,11 @@
+package sockets
+
+type GRPCRequest struct {
+}
+
+func (req *GRPCRequest) AddPayload(data []byte) {
+}
+
+func (req *GRPCRequest) String() string {
+	return ""
+}
