+++ internal/sockets/grpc_response.go
@@ -1,0 +1,11 @@
+package sockets
+
+type GRPCResponse struct {
+}
+
+func (resp *GRPCResponse) AddPayload(data []byte) {
+}
+
+func (resp *GRPCResponse) String() string {
+	return ""
+}
